requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis[hiredis]==5.2.1
pydantic-settings==2.7.0
bcrypt==4.2.1
PyJWT==2.10.1
geoip2==4.8.1
jinja2==3.1.5
python-multipart==0.0.20
